+++ packages/components/pages/f-account-info/test/component/f-accountInfo.component.spec.js
@@ -1,0 +1,14 @@
+const AccountInfo = require('../../test-utils/component-objects/f-accountInfo.component');
+
+const accountInfo = new AccountInfo();
+
+describe('f-accountInfo component tests', () => {
+    beforeEach(() => {
+        accountInfo.load();
+    });
+
+    it('should display the f-accountInfo component', () => {
+        // Assert
+        expect(accountInfo.isComponentDisplayed()).toBe(true);
+    });
+});