+++ packages/components/pages/f-checkout/src/components/Checkout.vue
@@ -1,23 +1,21 @@
 <template>
     <div>
-        <error-page
-            v-if="errorFormType"
-            :error-form-type="errorFormType"
-            :redirect-url="redirectUrl"
-            :service-type="serviceType" />
-
-        <age-verification
-            v-else-if="shouldShowAgeVerificationForm"
-            @checkout-verify-age="verifyCustomerAge" />
-
         <component
             :is="messageType.name"
-            v-else-if="message"
+            v-if="message && !errorFormType"
             ref="errorMessage"
             v-bind="messageType.props"
             @created="handleDialogCreation">
             <span>{{ messageType.content }}</span>
         </component>
+
+        <age-verification
+            v-else-if="shouldShowAgeVerificationForm"
+            @checkout-verify-age="verifyCustomerAge" />
+
+        <age-verification
+            v-else-if="shouldShowAgeVerificationForm"
+            @checkout-verify-age="verifyCustomerAge" />
 
         <div
             v-if="shouldShowCheckoutForm"
@@ -47,6 +45,12 @@
                 </template>
             </card>
         </div>
+
+        <error-page
+            v-else-if="errorFormType"
+            :error-form-type="errorFormType"
+            :redirect-url="redirectUrl"
+            :service-type="serviceType" />
     </div>
 </template>
 
@@ -207,8 +211,8 @@ export default {
     data () {
         return {
             tenantConfigs,
+            isLoading: true,
             errorFormType: null,
-            isFormSubmitting: false,
             availableFulfilmentTimesKey: 0,
             checkoutAnalyticsService: new CheckoutAnalyticsService(this)
         };
@@ -268,7 +272,7 @@ export default {
         },
 
         shouldShowCheckoutForm () {
-            return !this.errorFormType && !this.shouldShowAgeVerificationForm;
+            return !this.isLoading && !this.errorFormType && !this.shouldShowAgeVerificationForm;
         },
 
         shouldShowAgeVerificationForm () {
@@ -368,7 +372,6 @@ export default {
         await this.initialise();
         this.checkoutAnalyticsService.trackInitialLoad();
         this.$emit(EventNames.CheckoutMounted);
-        this.$parent.$emit(EventNames.StopLoadingSpinner); // We need to use `$this.$parent.$emit` here because checkout will be mounted as a slot in `f-spinner`.
     },
 
     methods: {
@@ -402,11 +405,15 @@ export default {
         async initialise () {
             this.setExperimentValues(this.experiments);
 
+            this.isLoading = true;
+
             const promises = this.isLoggedIn
                 ? [this.loadBasket(), this.loadCheckout(), this.loadAvailableFulfilment()]
                 : [this.loadBasket(), this.loadAddressFromLocalStorage(), this.loadAvailableFulfilment()];
 
             await Promise.all(promises);
+
+            this.resetLoadingState();
 
             if (this.shouldLoadAddress) {
                 await this.loadAddress();
@@ -796,6 +803,11 @@ export default {
             };
 
             this.handleEventLogging('CheckoutValidationError', validationState, { ...expandedData, validationState });
+        },
+
+        resetLoadingState () {
+            this.isLoading = false;
+            this.$parent.$emit(EventNames.StopLoadingSpinner); // We need to use `$this.$parent.$emit` here because checkout will be mounted as a slot in `f-spinner`.
         },
 
         getReferralState () {