+++ packages/components/pages/f-contact-preferences/src/services/mapping/contactPreferences.mapper.js
@@ -80,3 +80,4 @@ export const filterSortPreferences = preferences => {
 
     return model || [];
 };
+