+++ packages/components/pages/f-contact-preferences/stories/ContactPreferences.stories.js
@@ -1,12 +1,17 @@
+import Vue from 'vue';
+import Vuex from 'vuex';
 import { withA11y } from '@storybook/addon-a11y';
 import { select, withKnobs } from '@storybook/addon-knobs';
 import { locales } from '@justeat/storybook/constants/globalisation';
 import ContactPreferences from '../src/components/ContactPreferences.vue';
+import fContactPreferencesModule from '../src/store/contactPreferences.module';
 
 export default {
     title: 'Components/Pages',
     decorators: [withKnobs, withA11y]
 };
+
+Vue.use(Vuex);
 
 export const ContactPreferencesComponent = () => ({
     components: {
@@ -17,9 +22,14 @@ export const ContactPreferencesComponent = () => ({
             default: select('Locale', [locales.gb], locales.gb)
         }
     },
+    store: new Vuex.Store({
+        modules: {
+            fContactPreferencesModule
+        }
+    }),
     template: `<contact-preferences
         :locale="locale"
-        smart-gateway-base-url="https://localhost:8080" />`
+        smart-gateway-base-url="" />`
 });
 
 ContactPreferencesComponent.storyName = 'f-contact-preferences';