+++ packages/services/f-feature-management/src/vue/store/features.module.js
@@ -1,0 +1,19 @@
+import SET_FM_CONFIG_JSON from './mutation.types';
+
+export default {
+    state: () => ({
+        configJson: {}
+    }),
+
+    actions: {
+        update: ({ commit }, configJson) => {
+            commit(SET_FM_CONFIG_JSON, configJson);
+        }
+    },
+
+    mutations: {
+        [SET_FM_CONFIG_JSON]: (state, configJson) => {
+            state.configJson = configJson;
+        }
+    }
+};