+++ packages/components/pages/f-checkout/CHANGELOG.md
@@ -4,6 +4,22 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v3.7.0
+------------------------------
+*November 17, 2021*
+
+### Removed
+- Spinner from checkout form.
+
+
+v3.6.2
+------------------------------
+*November 17, 2021*
+
+### Fixed
+- Spinner and loading state added back in.
+
+
 v3.6.0
 ------------------------------
 *November 10, 2021*